backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
# python-magic==0.4.27  # Optional: for file type detection
# email parsing uses built-in email library

# Vector database and embeddings
pinecone-client==2.2.4
faiss-cpu==1.7.4
openai>=1.12.0
tiktoken==0.5.2

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Data processing and ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
aiofiles==23.2.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
